refactor: split CLI entry point from the generator engine
- src/cli.ts: commander wiring (watch/build/config commands),
  dependency-install caching, and process orchestration, extracted
  from src/index.ts.
- src/mdx-to-nextjs-generator.ts: the MDXToNextJSGenerator engine
  (scaffolding, watchers, page/layout regeneration), also extracted
  from src/index.ts.
- src/lib/generated-artifacts.ts: new GeneratedArtifacts class that
  tracks generated route and Markdown-mirror ownership in
  .doccupine-artifacts.json, migrating from the older
  .doccupine-llms-manifest.json / .doccupine-api-manifest.json files.
  Watch-mode renames and deletes now consult this registry instead of
  guessing routes from filenames, so a changed section removes only
  the files it owns.
- src/index.ts shrinks to a thin re-export of the split modules.
- Add matching test files for the new modules.
